# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChatCompletionRequest" : {
        "properties" : {
          "continuum_criticality" : {
            "type" : "string"
          },
          "continuum_deadline_ms" : {
            "format" : "int64",
            "type" : "integer"
          },
          "continuum_measure_uncertainty" : {
            "type" : "boolean"
          },
          "continuum_require_vision" : {
            "type" : "boolean"
          },
          "continuum_routing_mode" : {
            "type" : "string"
          },
          "max_completion_tokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "max_tokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "messages" : {
            "items" : {
              "$ref" : "#/components/schemas/ChatMessage"
            },
            "type" : "array"
          },
          "model" : {
            "type" : "string"
          },
          "n" : {
            "format" : "int32",
            "type" : "integer"
          },
          "response_format" : {
            "$ref" : "#/components/schemas/ResponseFormat"
          },
          "stop" : {
            "type" : "object"
          },
          "stream" : {
            "type" : "boolean"
          },
          "stream_options" : {
            "$ref" : "#/components/schemas/StreamOptions"
          },
          "temperature" : {
            "format" : "double",
            "type" : "number"
          },
          "tool_choice" : {
            "type" : "object"
          },
          "tools" : {
            "items" : {
              "$ref" : "#/components/schemas/Tool"
            },
            "type" : "array"
          },
          "top_p" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "ChatMessage" : {
        "properties" : {
          "content" : {
            "type" : "object"
          },
          "name" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "tool_call_id" : {
            "type" : "string"
          },
          "tool_calls" : {
            "items" : {
              "$ref" : "#/components/schemas/ToolCallDto"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "ChatRequest" : {
        "properties" : {
          "criticality" : {
            "type" : "string"
          },
          "deadlineMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "maxTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "measureUncertainty" : {
            "type" : "boolean"
          },
          "messages" : {
            "items" : {
              "$ref" : "#/components/schemas/Message"
            },
            "type" : "array"
          },
          "model" : {
            "type" : "string"
          },
          "requireVision" : {
            "type" : "boolean"
          },
          "responseFormat" : {
            "$ref" : "#/components/schemas/ResponseFormatRef"
          },
          "routingMode" : {
            "type" : "string"
          },
          "stream" : {
            "type" : "boolean"
          },
          "temperature" : {
            "format" : "double",
            "type" : "number"
          },
          "toolChoice" : {
            "type" : "string"
          },
          "tools" : {
            "items" : {
              "$ref" : "#/components/schemas/ToolRef"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "ChatResponse" : {
        "properties" : {
          "agreementClusters" : {
            "format" : "int32",
            "type" : "integer"
          },
          "choices" : {
            "items" : {
              "additionalProperties" : {
                "type" : "object"
              },
              "type" : "object"
            },
            "type" : "array"
          },
          "completionTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "confidence" : {
            "format" : "double",
            "type" : "number"
          },
          "cost" : {
            "format" : "double",
            "type" : "number"
          },
          "failovers" : {
            "format" : "int32",
            "type" : "integer"
          },
          "finishReason" : {
            "type" : "string"
          },
          "latency" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lowConfidence" : {
            "type" : "boolean"
          },
          "model" : {
            "type" : "string"
          },
          "promptTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "provider" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          },
          "response" : {
            "type" : "string"
          },
          "routingReason" : {
            "type" : "string"
          },
          "tokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "toolCalls" : {
            "items" : {
              "$ref" : "#/components/schemas/ToolCallRef"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "FunctionCall" : {
        "properties" : {
          "arguments" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "FunctionDef" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "parameters" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "ImageRef" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Message" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "images" : {
            "items" : {
              "$ref" : "#/components/schemas/ImageRef"
            },
            "type" : "array"
          },
          "role" : {
            "type" : "string"
          },
          "toolCallId" : {
            "type" : "string"
          },
          "toolCalls" : {
            "items" : {
              "$ref" : "#/components/schemas/ToolCallRef"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "ResponseFormat" : {
        "properties" : {
          "json_schema" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ResponseFormatRef" : {
        "properties" : {
          "schema" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "StreamOptions" : {
        "properties" : {
          "include_usage" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "Tool" : {
        "properties" : {
          "function" : {
            "$ref" : "#/components/schemas/FunctionDef"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ToolCallDto" : {
        "properties" : {
          "function" : {
            "$ref" : "#/components/schemas/FunctionCall"
          },
          "id" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ToolCallRef" : {
        "properties" : {
          "argumentsJson" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ToolRef" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "parameters" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Durable execution engine for AI agents: workflows, gateway and console APIs.",
    "title" : "Continuum",
    "version" : "1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/portal/developer/playground" : {
      "post" : {
        "operationId" : "playground",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChatRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChatResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "portal-developer-controller" ]
      }
    },
...
    "/v1/chat/completions" : {
      "post" : {
        "operationId" : "chatCompletions",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "open-ai-compat-controller" ]
      }
    },
...
  }
}